--- src/components/Filter/Filter.module.scss
@@ -1,74 +1,0 @@
-.filter {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-bottom: 51px;
-  &__title {
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    margin-right: 15px;
-  }
-
-  &__choice {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-
-  &__menu {
-    position: absolute;
-    margin-top: 50px;
-    width: 248px;
-    height: 305px;
-    background-color: #313131;
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-    border-radius: 5%;
-    overflow-y: scroll;
-    border: 20px solid transparent;
-
-    &::-webkit-scrollbar {
-      width: 5px;
-      background-color: #202020;
-    }
-    &::-webkit-scrollbar-thumb {
-      background-color: #ffffff;
-    }
-  }
-  &__item {
-    display: flex;
-    flex-direction: column;
-  }
-
-  &__link {
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    color: white;
-
-    cursor: pointer;
-
-    &:hover {
-      color: #ad61ff;
-      text-decoration-line: underline;
-    }
-  }
-
-  &__button {
-    font-style: normal;
-    font-weight: 400;
-    font-size: 16px;
-    line-height: 24px;
-    border: 1px solid #ffffff;
-    border-radius: 60px;
-    padding: 6px 20px;
-
-    &:not(:last-child) {
-      margin-right: 10px;
-    }
-  }
-}
