+++ src/components/Nav/Nav.module.css
@@ -1,0 +1,51 @@
+.nav {
+  width: 244px;
+  background-color: #181818;
+  padding: 20px 0 20px 36px;
+}
+
+.logo {
+  width: 113.33px;
+  height: 43px;
+  padding: 13px 0 13px 0;
+  background-color: transparent;
+  margin-bottom: 20px;
+}
+.logo__image {
+  width: 113.33px;
+  height: 17px;
+  color: #181818;
+}
+
+.burger {
+  width: 20px;
+  height: 36px;
+  padding: 13px 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+.burger__line {
+  display: inline-block;
+  width: 100%;
+  height: 1px;
+  background-color: #D3D3D3;
+}
+
+.menu {
+  display: block;
+  visibility: visible;
+}
+.menu__list {
+  padding: 18px 0 10px 0;
+}
+.menu__item {
+  padding: 5px 0;
+  margin-bottom: 16px;
+}
+.menu__link {
+  color: #FFFFFF;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+}/*# sourceMappingURL=Nav.module.css.map */
