+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/client';
 
 import { store } from './redux/slices/store';
 
-
+import './style.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
