+++ src/pages/home/home.module.scss
@@ -223,7 +223,66 @@
 
 
 
+.sidebar {
 
+    &__personal {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        justify-content: flex-end;
+        padding: 12px 0 15px 0;
+    }
+
+    &__personal_name {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+        color: #FFFFFF;
+        margin-right: 16px;
+    }
+
+    &__avatar {
+        width: 43px;
+        height: 43px;
+        background-color: #313131;
+        border-radius: 50%;
+    }
+
+    &__block {
+        height: 100%;
+        padding: 240px 0 0 0;
+        display: flex;
+        flex-direction: column;
+        justify-content: flex-start;
+        margin-right: 150px;
+    }
+
+    &__list {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+
+    &__item {
+        width: 250px;
+        height: 150px;
+
+        &:not(:last-child) {
+            margin-bottom: 30px;
+        }
+    }
+
+    &__link {
+        width: 100%;
+        height: 100%;
+    }
+
+    &__img {
+        width: 100%;
+        height: auto;
+    }
+}
 
 
 .centerblock {
@@ -247,6 +306,81 @@
         width: 1136px;
         height: 100%;
         margin-bottom: 50px;
+    }
+}
+
+
+.filter {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-bottom: 51px;
+    &__title {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+        margin-right: 15px;
+    }
+
+    &__choice {
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+    }
+
+    &__menu {
+        position: absolute;
+        margin-top: 50px;
+        width: 248px;
+        height: 305px;
+        background-color: #313131;
+        display: flex;
+        flex-direction: column;
+        gap: 15px;
+        border-radius: 5%;
+        overflow-y: scroll;
+        border: 20px solid transparent;
+
+        &::-webkit-scrollbar {
+            width: 5px;
+            background-color: #202020;
+        }
+        &::-webkit-scrollbar-thumb {
+            background-color: #FFFFFF;
+        }
+
+    }
+
+    
+
+    &__link {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+        color: white;
+        
+        cursor: pointer;
+
+        &:hover {
+            color: #AD61FF;
+            text-decoration-line: underline;
+        }
+    }
+
+    &__button {
+        font-style: normal;
+        font-weight: 400;
+        font-size: 16px;
+        line-height: 24px;
+        border: 1px solid #FFFFFF;
+        border-radius: 60px;
+        padding: 6px 20px;
+
+        &:not(:last-child) {
+            margin-right: 10px;
+        }
     }
 }
 
@@ -285,6 +419,7 @@
         stroke: #696969;
     }
 }
+
 .col01 {
     width: 447px;
 }
