+++ src/redux/slices/store.js
@@ -1,7 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import song from './song';
 
-
 export const store = configureStore({
   reducer: {
     song,
