+++ src/style.module.css
@@ -1,0 +1,281 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+*:before,
+*:after {
+  box-sizing: border-box;
+}
+
+a,
+a:visited {
+  text-decoration: none;
+  font-family: "Stratos-Regular", sans-serif;
+  cursor: pointer;
+}
+
+button,
+.btn {
+  cursor: pointer;
+}
+
+ul li {
+  list-style: none;
+  text-decoration: none;
+}
+
+@font-face {
+  font-family: "Stratos-Regular";
+  src: local("Stratos-Regular"), local("Stratos-Regular"), url("./fonts/example stratos font/Stratos-Regular.woff2") format("woff2"), url("./fonts/example stratos font/Stratos-Regular.woff") format("woff");
+  font-weight: 400;
+  font-style: normal;
+}
+html,
+body {
+  width: 100%;
+  height: 100%;
+  font-family: "Stratos-Regular", sans-serif;
+  color: #FFFFFF;
+}
+
+.wrapper {
+  width: 100%;
+  min-height: 100%;
+  overflow: hidden;
+  background-color: #383838;
+}
+
+.container {
+  max-width: 1920px;
+  height: 100vh;
+  margin: 0 auto;
+  position: relative;
+  background-color: #181818;
+}
+
+.btn_text:hover {
+  border-color: #D9B6FF;
+  color: #D9B6FF;
+  cursor: pointer;
+}
+
+.btn_icon:hover svg {
+  fill: transparent;
+  stroke: #ACACAC;
+  cursor: pointer;
+}
+
+.btn_text:active {
+  border-color: #AD61FF;
+  color: #AD61FF;
+  cursor: pointer;
+}
+
+.btn_icon:active svg {
+  fill: transparent;
+  stroke: #FFFFFF;
+  cursor: pointer;
+}
+
+.btn_icon:active .track-play__like-svg,
+.btn_icon:active .track-play__dislike-svg {
+  fill: #696969;
+  stroke: #FFFFFF;
+  cursor: pointer;
+}
+
+.main {
+  flex: 1 1 auto;
+  display: flex;
+  justify-content: space-between;
+}
+.main__nav {
+  width: 244px;
+  background-color: #181818;
+  padding: 20px 0 20px 36px;
+}
+.main__sidebar {
+  max-width: 418px;
+  padding: 20px 90px 20px 78px;
+}
+
+.centerblock {
+  width: auto;
+  flex-grow: 3;
+  padding: 20px 40px 20px 111px;
+}
+.centerblock__h2 {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 64px;
+  line-height: 72px;
+  letter-spacing: -0.8px;
+  margin-bottom: 45px;
+}
+.centerblock__content {
+  display: flex;
+  flex-direction: column;
+  width: 1136px;
+  height: 100%;
+}
+
+.filter {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-bottom: 51px;
+}
+.filter__title {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  margin-right: 15px;
+}
+.filter__choice {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+.filter__menu {
+  position: absolute;
+  margin-top: 50px;
+  width: 248px;
+  height: 305px;
+  background-color: #313131;
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+  border-radius: 5%;
+  overflow-y: scroll;
+  border: 20px solid transparent;
+}
+.filter__menu::-webkit-scrollbar {
+  width: 5px;
+  background-color: #202020;
+}
+.filter__menu::-webkit-scrollbar-thumb {
+  background-color: #FFFFFF;
+}
+.filter__link {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  color: white;
+  cursor: pointer;
+}
+.filter__link:hover {
+  color: #AD61FF;
+  text-decoration-line: underline;
+}
+.filter__button {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  border: 1px solid #FFFFFF;
+  border-radius: 60px;
+  padding: 6px 20px;
+}
+.filter__button:not(:last-child) {
+  margin-right: 10px;
+}
+
+.content__playlist {
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+}
+
+.playlist_title {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 24px;
+}
+.playlist_title__col {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 24px;
+  letter-spacing: 2px;
+  color: #696969;
+  text-transform: uppercase;
+}
+.playlist_title__svg {
+  width: 12px;
+  height: 12px;
+  fill: transparent;
+  stroke: #696969;
+}
+
+.col01 {
+  width: 447px;
+}
+
+.col02 {
+  width: 321px;
+}
+
+.col03 {
+  width: 245px;
+}
+
+.col04 {
+  width: 60px;
+  text-align: end;
+}
+
+.sidebar__personal {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-end;
+  padding: 12px 0 15px 0;
+}
+.sidebar__personal_name {
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  color: #FFFFFF;
+  margin-right: 16px;
+}
+.sidebar__avatar {
+  width: 43px;
+  height: 43px;
+  background-color: #313131;
+  border-radius: 50%;
+}
+.sidebar__block {
+  height: 100%;
+  padding: 240px 0 0 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  margin-right: 150px;
+}
+.sidebar__list {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.sidebar__item {
+  width: 250px;
+  height: 150px;
+}
+.sidebar__item:not(:last-child) {
+  margin-bottom: 30px;
+}
+.sidebar__link {
+  width: 100%;
+  height: 100%;
+}
+.sidebar__img {
+  width: 100%;
+  height: auto;
+}/*# sourceMappingURL=style.module.css.map */
